Fix for bug #25421: MySQL threads don't respond to the kill command
Calculating the estimated number of records for a range scan may take a
significant time, and it was impossible for a user to interrupt that
process by killing the connection or the query.

Fixed by checking the thread's 'killed' status in check_quick_keys() and
interrupting the calculation process if it is set to a non-zero value.


sql/opt_range.cc:
  Check the thread's 'killed' status in check_quick_keys() and
  interrupt the calculation process if it is set to a non-zero value.