[PATCH] WL4081 Allow use of direct IO (O_DIRECT) with azio
Change azio to either posix_memalign allocate in and out buffers OR
allow user to allocate these. Flag in azio_stream should be where
padding was so no additional memory used.

This enables O_DIRECT usage with azio if aligned buffers are allocated
(by default, the azio allocation - posix_memalign - does this).

Also fiddles with ZLIB internals to get the correct buffer sizes to allocate
for deflate and inflate.