Bug#32948: FKs allowed to reference partitioned table
Problem: when alter to partitioned table,
it does not see it as change of engine.

Solution: If alter includes partitioning, check if it is possible
to change engines (eg. is the table referenced by a FK)


mysql-test/r/partition_innodb.result:
  Bug#32948: FKs allowed to reference partitioned table
  test result
mysql-test/t/partition_innodb.test:
  Bug#32948: FKs allowed to reference partitioned table
  test case
sql/sql_table.cc:
  Bug#32948: FKs allowed to reference partitioned table
  
  if alter to partitioning, it is the same as changing engine.