Bug #33332 assert at the end of mysql_binlog_send
The heartbeat code did not check possible error in net_flush which
kept the dump thread hanging in heartbeating loop even though
the slave has disappeared.
When eventaully the dump thread was killed its cleanup did not use the error
path and that led to an assert.

Fixed with adding the check of net_flush return value. If one happens
send_heartbeat_event() returns the error which forces the dump thread
to exit along the error path.