bug#32003 rpl.rpl_server_id2 can cause extraneous errors recorded in error log
the error printed in the slave error log appeared due to slave's threads
were killed in the middle of handling START SLAVE.
The sql command is asynchronous so that START may not be complete by
the time a test issues STOP.
Clearly, printing the error is not necessary if the threads are killed.

Fixed with deploying a condition for printing the error that previously
was introduced by bug#29976.