Bug#24624: MySQL client ignores its own completion table
Nothing was in the command-line dictionary, because of the wrong 
order of instructions when populating it.

This is a smaller, less optimistic patch that both fixes a bug and
refreshes the list of keywords that the command-line library (e.g., 
readline) can use to expand typed commands.  Now, read from the 
command list /after/ we free the list, not before.

The best way is to read the keywords from the lexer code, but that
doesn't work everywhere yet.  Grr.