Bug#31331 - MyISAM or Merge Table upgrade incompatibility with 5.1
A table with BLOB/TEXT prefix key part, created with version 4.1,
could not be opened by a 5.1 server.

The routine check at table open, if the frm file matches the MyISAM
table, was too picky regarding old and new implementation of such
keys.

Added relaxed check for blob prefix key part.

No test case. It requires to create a table in 4.1 and open it in
5.1.