bug#34427 slave misses rendezvous in rpl_variables
There was no instruction in the test that enforces the slave successfully connect
to the master.
The way the test was been written allowed the slave to had been late for rendezvous 
so that about-connecting time queries to the master failed and are error-logged
to had been seen in Warnings of pb.

Fixed with adding a sychronization primitive to the test.
No test case is possible, observe error logs on pb.

Todo: revise need of rpl_report.pl's rules due to failing execution of
queries from get_master_verion_and_clock().
Any test should try to use a synchornization primitive like the current fix
makes and do not let the slave to miss successful connecting.