Bug #34458  	Extreneous use of templates in server code
Replacing a template function with a normal static function.
The template parameter, which previously was the class to
find a binlogging function in, is now passed as a pointer to
the actual binlogging function instead.

The patch requires change of indention, but that is submitted
as a separate patch.


sql/handler.cc:
  Replacing template function with a normal static function.