Bug#34726: open_tables() crashes server if running with --debug
The DBUG code emits the current value of the proc_info member of THD,
which may be set to NULL.  It was wrong to dereference that value
with the format string %s without verifying that it was valid.

Now, insert an inline test that substitutes the string "(null)" for
NULL pointers.