In bug#31233 mysql_alter_table() fails to drop UNIQUE KEY it was descovered that the online add/drop index implementation does not handle unique indexes that are promoted to primary keys. In MySQL, if a table has no primary key the first unique index defined over non-nullable columns is promoted to become the primary key. Two special cases must be handled:
1. If a table has no primary key and a unique index defined over non-nullable
columns is added, only storage engines that support adding primary keys online
should be called.
2. If a table has no explicit primary key, but a unique index defined over non-nullable
columns (promoted as implicit primary key) and this index is dropped, only storage engines
that support dropping primary keys should be called.

To detect these cases the function compare_tables was modified to count
all candidate_keys (keys that are possible to promote to primary keys) of a table
to detect:
 1. candidate_key_count == 0, and
 2. candidate_key_count == 1