branches/zip: Remove innodb.patch.
This patch tries to enable resizeable buffer pool by polling the configuration
parameter for the buffer pool size, which is not a good solution.  The
right way could be to have an update function callback of a settable
MySQL variable to send an event to the master thread.

It could also make sense to expose the buffer pool chunks to the user.
The first chunk would never be freed.  Other chunks than the first one
would only be used for allocating page frames (uncompressed or compressed)
and block descriptors of of compressed pages (buf_page_t).  That is, other
users of the buffer pool, such as mem_heap_create_block() and the lock
table, would be restricted to the first chunk.  This would allow other
chunks to be freed by simply flushing any dirty blocks that they contain.

It might also be worthwhile to create multiple chunks initially, based on
the initial buffer pool size and the HugeTLB page size.  In that way, the
buffer pool could be reduced from the initial configuration at runtime.