Bug#30378: mysql.server needs to skip wait_for_pid() when mysqld \ 	isn't running
Pass the process id of the manager as a parameter to "wait_for_pid"
and if the manager isn't running, then do not continue to wait.

Also, capture the error message of our process-existence test, 
"kill -0", as we expect errors and shouldn't pass them to the user.  
Additionally, be a bit more descriptive of what the problem is.