Fix for Bug#34274: Invalid handling of 'DEFAULT 0' for YEAR data type.
The problem was that for some unknown reason 0 was not allowed
as a default value for YEAR data type. That was coded before BK.
However the Manual does not say a word about such a limitation.
Also, it looks inconsistent with other data types.

The fix is to allow 0 as a default value.


mysql-test/r/create.result:
  Update result file.
mysql-test/t/create.test:
  Add a test case for Bug#34274: Invalid handling of 'DEFAULT 0'
  for YEAR data type.
sql/unireg.cc:
  Allow 0 as a default value for YEAR data type.